Fixed bug in 'count-9s-from-1-to-n'
Changed a manipulation of the index i of the for loop so that it dynamically changes
with the place value of the number for each iteration. Before the fix, if several combined zeros
occured in the number, the place value and index i got out of sync and the calculations became errenous.